deprecated the old ReactiveMap component
added textbox functionality
added tooltip for textbox
