+++ package.json
@@ -35,6 +35,7 @@
     "react-router-dom": "^5.1.0",
     "react-select": "^2.4.4",
     "react-table": "^6.10.3",
+    "react-tooltip": "^3.11.1",
     "style-loader": "^0.23.1",
     "url-loader": "^1.1.2",
     "webpack": "^4.41.0",
