+++ src/components/ReactiveMap/index.jsx
@@ -1,279 +1,276 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { ReactiveComponent } from '@appbaseio/reactivesearch';
-import { Map, TileLayer, FeatureGroup, Polygon, ImageOverlay, Popup, Marker } from 'react-leaflet';
-import { EditControl } from "react-leaflet-draw";
-
-import { LEAFLET_TILELAYER, LEAFLET_ATTRIBUTION, DEFAULT_MAP_DISPLAY } from '../../config.js';
-import { _extractPolygonData, _switchPolygonCoordinates } from './utils.js'
+import ReactTooltip from 'react-tooltip'
+import L from 'leaflet';
+import 'leaflet-draw';
 
 import 'leaflet/dist/leaflet.css';
 import 'leaflet-draw/dist/leaflet.draw.css';
 import './style.css';
 
-import L from 'leaflet';
-
-delete L.Icon.Default.prototype._getIconUrl;
-
-L.Icon.Default.mergeOptions({
-  iconRetinaUrl: require('leaflet/dist/images/marker-icon-2x.png'),
-  iconUrl: require('leaflet/dist/images/marker-icon.png'),
-  shadowUrl: require('leaflet/dist/images/marker-shadow.png')
-});
+import { LEAFLET_TILELAYER, LEAFLET_ATTRIBUTION, DEFAULT_MAP_DISPLAY } from '../../config.js';
 
 const DRAW_POLYGON_COLOR = '#f06eaa';
 const DRAW_POLYGON_WEIGHT = 7;
 
-/**
- * Map in Tosca to trigger ES searches based on geo location drawing
- * TODO:
- *    migrate data from GRQ to ES 7 (KIND OF DONE)
- *    make es query compatible with ReactiveSearch ** (DONE)
- *    collapsable text field on the bottom for coordinates (DONE I THINK)
- *    add images to the map (DONE)
- *    add highlighting when clicking (need callbacks)
- */
-var MapComponent = class extends React.Component {
+let MapComponent = class extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
       displayMap: DEFAULT_MAP_DISPLAY,
-      textBoxValue: '',
-      zoomLevel: props.defaultZoom
+      value: null,
+      polygonTextbox: props.value,
     };
-
-    this.toggleMapDisplay = this.toggleMapDisplay.bind(this);
-    this.polygonCreateEvent = this.polygonCreateEvent.bind(this);
-    this.removePolygons = this.removePolygons.bind(this);
-    this.polygonTextInput = this.polygonTextInput.bind(this);
-    this.polygonTextChange = this.polygonTextChange.bind(this);
+    this._toggleMapDisplay = this._toggleMapDisplay.bind(this);
+    this._polygonTextChange = this._polygonTextChange.bind(this);
+    this._polygonTextInput = this._polygonTextInput.bind(this);
   }
 
   componentDidMount() {
-    const { setPolygon } = this.props;
-    if (setPolygon) {
-      const polyArr = JSON.parse(setPolygon);
-      this._setQuery(this.props, polyArr);
+    const that = this; // so we can run ReactiveComponent methods
 
-      let inputPolygon = (setPolygon === null || setPolygon.length === 0) ? null : JSON.parse(setPolygon);
+    this.map = L.map('leaflet-map-id', { // initializing the map
+      center: [36.7783, -119.4179],
+      maxZoom: this.props.maxZoom,
+      minZoom: this.props.minZoom,
+      zoom: localStorage.getItem('zoom') || this.props.zoom,
+      layers: [L.tileLayer(LEAFLET_TILELAYER, { attribution: LEAFLET_ATTRIBUTION })]
+    });
+
+    this.drawnItems = new L.FeatureGroup();
+    this.map.addLayer(this.drawnItems);
+
+    this.layerGroup = L.layerGroup().addTo(this.map);  // store all dataset panels in this layergroup
+
+    let drawControl = new L.Control.Draw({
+      edit: {
+        featureGroup: this.drawnItems,
+        remove: false,
+      },
+      draw: {
+        circle: false,
+        marker: false,
+        polyline: false,
+        circlemarker: false,
+        polygon: {
+          shapeOptions: {
+            color: DRAW_POLYGON_COLOR,
+            weight: DRAW_POLYGON_WEIGHT
+          }
+        },
+        rectangle: {
+          shapeOptions: {
+            color: DRAW_POLYGON_COLOR,
+            weight: DRAW_POLYGON_WEIGHT
+          }
+        },
+      },
+    });
+    this.map.addControl(drawControl);
+
+    this.map.on('zoomend', this._zoomHandler);
+    this.map.on('draw:drawstart', this._clearDatasets);
+    this.map.on('draw:created', (event) => {
+      let newLayer = event.layer;
+      newLayer.options = {
+        ...newLayer.options,
+        ...{ weight: DRAW_POLYGON_WEIGHT, color: DRAW_POLYGON_COLOR }
+      };
+      let polygon = newLayer.getLatLngs()[0].map(cord => [cord.lng, cord.lat]);
+      polygon = [...polygon, polygon[0]];
+      this.drawnItems.addLayer(newLayer, false);
+
+      const query = this._generateQuery(polygon);
+      const polygonString = JSON.stringify(polygon);
+
+      that.props.setQuery({
+        query,
+        value: polygonString
+      });
+      that.setState({
+        value: polygonString,
+        polygonTextbox: polygonString,
+      });
+    });
+  }
+
+  componentDidUpdate() {
+    if (this.props.value !== this.state.value) { // prevent maximum recursion error
+      if (this.props.value !== null) { // if the page loads with coordinates in the URL
+        let polygon = JSON.parse(this.props.value);
+        const query = this._generateQuery(polygon);
+        this.props.setQuery({
+          query,
+          value: this.props.value
+        });
+      } else { // handles onClear (facets)
+        this.sendEmptyQuery();
+      }
+
       this.setState({
-        inputPolygon: _switchPolygonCoordinates(inputPolygon),
-        textBoxValue: JSON.stringify(inputPolygon)
+        value: this.props.value,  // prevent maximum recursion error
+        polygonTextbox: this.props.value
       });
     }
   }
 
-  componentDidUpdate(prevProps, prevState) {
-    // when they clear the facet so it doesnt query with empty coordinates
-    if (!this.props.setPolygon) {
-      this.props.setQuery({
-        query: null,
-        value: [], // don't know why it works but as long as query is null we're good
-      });
-
-      this.setState({
-        inputPolygon: [], // so we can remove the defualt polygon when we "clear all"
-        textBoxValue: ''
-      });
-
-      this.removePolygons();
-    }
-  }
-
-  toggleMapDisplay() {
-    const { map } = this.refs;
+  sendEmptyQuery = () => { // remove the bbox facet
+    this.drawnItems.clearLayers();
+    this.props.setQuery({
+      query: null,
+      value: null
+    });
     this.setState({
-      displayMap: !this.state.displayMap
-    },
-      map.leafletElement._onResize() // this forces leaflet to redraw the map
-    );
+      polygonTextbox: null
+    });
   }
 
-  _setQuery(props, polygon) { // passing the ES query back to the wrapper: ReactiveComponent
-    const query = {
-      query: {
-        "bool": {
-          "filter": {
-            "geo_shape": {
-              "location": {
-                "shape": {
-                  "type": "polygon",
-                  "coordinates": [polygon]
-                }
+  // client side event handlers
+  _clearDatasets = () => this.drawnItems.clearLayers();
+  _zoomHandler = () => localStorage.setItem('zoom', this.map.getZoom());
+  _reRenderMap = () => this.map._onResize();
+  _toggleMapDisplay = () => this.setState({ displayMap: !this.state.displayMap }, this._reRenderMap);
+  _polygonTextChange = (e) => {
+    this.setState({
+      polygonTextbox: e.target.value
+    });
+  }
+  _polygonTextInput = (e) => {
+    if (e.key === 'Enter' && e.shiftKey) {
+      e.preventDefault();
+      try {
+        const polygonString = e.target.value;
+        if (polygonString.trim() === '') {
+          this.sendEmptyQuery();
+          return;
+        };
+
+        let polygon = JSON.parse(polygonString);
+        const query = this._generateQuery(polygon);
+
+        this.props.setQuery({
+          query,
+          value: polygonString
+        });
+        this.setState({
+          value: polygonString,
+          polygonTextbox: polygonString,
+        });
+      } catch (err) {
+        alert('Not valid JSON');
+      }
+    }
+  }
+
+  // utility function to handle the data
+  _generateQuery = (polygon) => ({
+    query: {
+      "bool": {
+        "filter": {
+          "geo_shape": {
+            "location": {
+              "shape": {
+                "type": "polygon",
+                "coordinates": [polygon]
               }
             }
           }
         }
       }
-    };
-    props.setQuery({
-      query,
-      value: JSON.stringify(polygon)
-    })
+    }
+  });
+  _switchCoordinates = (polygon) => (polygon.map((row) => [row[1], row[0]]));
+  _transformData = (data) => { // transforms data for map
+    const displayData = data.map(row => ({
+      _id: row._id,
+      _index: row._index,
+      key: `${row._index}/${row._id}`,
+      coordinates: this._switchCoordinates(row.location.coordinates[0]),
+      center: [row.center.coordinates[1], row.center.coordinates[0]],
+      image: `${row.urls[0]}/${row._id}.interferogram.browse_coarse.png`
+    }));
+    return displayData
   }
 
-  polygonCreateEvent(e) {
-    const polygon = e.layer.getLatLngs()[0].map(cord => [cord.lng, cord.lat]);
-    polygon.push(polygon[0]); // GEOJSON NEED THE LAST POINT TO BE THE SAME AS THE FIRST POINT
-    this._setQuery(this.props, polygon);
+  _renderBbox = () => {
+    // rendering pink bbox if not drawn and specified in facets (for page on load or backwards forwards on browser)
+    const { value } = this.props;
+    const drawnItems = this.drawnItems;
 
-    this.setState({
-      inputPolygon: null,
-      textBoxValue: JSON.stringify(polygon)
-    });
-  }
-
-  polygonTextChange(e) {
-    this.setState({
-      textBoxValue: e.target.value
-    });
-  }
-
-  polygonTextInput(e) {
-    if (e.key === 'Enter' && e.shiftKey) {
-      try {
-        const textValue = e.target.value;
-        if (textValue.trim() === '') return;  // do nothing if text is blank
-
-        let inputPolygon = JSON.parse(textValue);
-        this.setState({
-          inputPolygon: _switchPolygonCoordinates(inputPolygon),
-          textBoxValue: JSON.stringify(inputPolygon)
+    if (drawnItems) {
+      drawnItems.clearLayers();
+      if (value) { // manually draw bounding box
+        let coordinates = this._switchCoordinates(JSON.parse(value));
+        let poly = L.polygon(coordinates, {
+          color: DRAW_POLYGON_COLOR,
+          weight: DRAW_POLYGON_WEIGHT,
+          opacity: 0.5,
         });
-
-        this.removePolygons();
-        this._setQuery(this.props, inputPolygon);
-      } catch (err) {
-        alert("Not valid JSON");
+        poly.addTo(drawnItems).addTo(this.map);
       }
-      e.preventDefault();
     }
   }
 
-  removePolygons() { // to remove the already drawn polygon if you're drawing a new one
-    const { edit } = this.refs;
-    const layerContainer = edit.leafletElement.options.edit.featureGroup;
-    const layers = layerContainer._layers;
-    const layerIds = Object.keys(layers);
+  _renderDatasets = () => {
+    const { data } = this.props;
+    const layerGroup = this.layerGroup;
 
-    layerIds.forEach(id => {
-      var layer = layers[id];
-      layerContainer.removeLayer(layer);
-    });
+    if (layerGroup) {
+      layerGroup.clearLayers(); // clearing all the previous datasets
+      this._transformData(data).map(row => { // parsing data and rendering map
+        let poly = L.polygon(row.coordinates, {
+          fillOpacity: 0,
+          weight: 1.3
+        });
+        poly.bindPopup(row._id).addTo(layerGroup).addTo(this.map);
+      });
+    }
   }
 
-  render() {
-    const { defaultZoom, maxZoom, minZoom, facetData } = this.props;
-    const { displayMap, inputPolygon, textBoxValue, zoomLevel } = this.state;
-    const mapContainerStyles = displayMap ? {} : { display: 'none' }; // to toggle the map
 
-    // extracting and morphing the coordinates in the data from elasticsearch
-    const extractedData = _extractPolygonData(facetData);
-    const center = extractedData.length === 0 ? [36.7783, -119.4179] : extractedData[0].center;
+  render() {
+    const { data, value } = this.props;
+    const { displayMap, polygonTextbox } = this.state;
+
+    this._renderBbox(); // rendering pink bbox
+    this._renderDatasets(); // rendering dataset panes
+
+    // find first occurance of valid center coordinate
+    let validCenter = data.find(row => row.center.coordinates)
+    if (validCenter) {
+      const centerCoordinates = validCenter.center.coordinates;
+      this.map.panTo(new L.LatLng(centerCoordinates[1], centerCoordinates[0]));
+    }
+
+    const textboxTooltip = `Press SHIFT + ENTER to manually input polygon... ex. [ [-125.09335, 42.47589], ... ,[-125.09335, 42.47589] ]`;
+    const mapStyle = {
+      width: '95%',
+      height: '59em',
+      margin: '0 auto',
+      marginBottom: '15px',
+      display: displayMap ? 'block' : 'none'
+    };
 
     return (
       <div className="reactive-map-container">
-        <button
-          onClick={this.toggleMapDisplay}>
+        <button className="toggle-map-btn" onClick={this._toggleMapDisplay}>
           {displayMap ? 'Hide Map' : 'Show Map'}
         </button>
-        <div className='map-container'>
-          <Map
-            ref='map'
-            center={center}
-            zoom={zoomLevel}
-            minZoom={minZoom}
-            maxZoom={maxZoom}
-            style={mapContainerStyles}
-          // onViewportChange={e => this.setState({ zoomLevel: e.zoom })}
-          >
-            <TileLayer
-              url={LEAFLET_TILELAYER}
-              attribution={LEAFLET_ATTRIBUTION}
-            />
-            <FeatureGroup>
-              <EditControl
-                ref="edit"
-                position='topleft'
-                onDrawStart={this.removePolygons}
-                onCreated={this.polygonCreateEvent}
-                onDeleted={this._onDeleted}
-                draw={{
-                  circle: false,
-                  marker: false,
-                  polyline: false,
-                  circlemarker: false,
-                  polygon: {
-                    shapeOptions: {
-                      color: DRAW_POLYGON_COLOR,
-                      weight: DRAW_POLYGON_WEIGHT,
-                      fill: false,
-                    }
-                  },
-                  rectangle: {
-                    shapeOptions: {
-                      color: DRAW_POLYGON_COLOR,
-                      weight: DRAW_POLYGON_WEIGHT,
-                      fill: false,
-                    }
-                  },
-                }}
-                edit={{
-                  remove: false,
-                  edit: false
-                }}
-              />
-            </FeatureGroup>
-            {
-              inputPolygon !== [] && inputPolygon !== null && inputPolygon !== undefined ?
-                (<Polygon positions={inputPolygon} color="#3388ff" opacity={0.5} fillOpacity={0.2} />) :
-                null
-            }
-            {extractedData.map(row => (
-              <Fragment>
-                <Polygon
-                  key={`${row._index}/${row._id}`}
-                  positions={row.polygon}
-                  color="#3388ff"
-                  opacity={1}
-                  fillOpacity={0}
-                  weight={1.1}
-                >
-                  <Popup>
-                    <div style={{ wordWrap: 'break-word' }} onClick={this.props.clickIdHandler.bind(this, row._id)}>
-                      {row._id}
-                    </div>
-                  </Popup>
-                </Polygon>
-                {/* <Marker position={row.center}>
-                  <Popup>
-                    <div style={{ wordWrap: 'break-word' }} onClick={this.props.clickIdHandler.bind(this, row._id)}>
-                      {row._id}
-                    </div>
-                  </Popup>
-                </Marker> */}
-              </Fragment>
-            ))}
-            {extractedData.map(row => (
-              <ImageOverlay
-                key={`${row._index}/${row._id}_imageOverlay`}
-                url={row.imageUrl}
-                bounds={row.polygon}
-                opacity={1}
-              />
-            ))}
-          </Map>
-        </div>
+        <div id="leaflet-map-id" style={mapStyle} />
+
+        <ReactTooltip place="top" type="dark" effect="solid" />
         <textarea
-          className='map-coordinates-textbox'
-          placeholder={`Press SHIFT + ENTER to manually input polygon...\n ex. [ [-125.09335, 42.47589], ... ,[-125.09335, 42.47589] ]`}
-          value={textBoxValue}
-          onChange={this.polygonTextChange}
-          onKeyPress={this.polygonTextInput}
+          className="map-coordinates-textbox"
+          placeholder={textboxTooltip}
+          data-tip={textboxTooltip}
+          value={polygonTextbox || ''}
+          onChange={this._polygonTextChange}
+          onKeyPress={this._polygonTextInput}
         >
         </textarea>
       </div>
-    )
+    );
   }
 }
 
@@ -284,20 +281,18 @@ export default class ReactiveMap extends React.Component {
   }
 
   render() {
+    const { componentId, data } = this.props;
+
     return (
       <ReactiveComponent
-        componentId={this.props.mapComponentId}
+        componentId={componentId}
         URLParams={true}
         render={({ setQuery, value }) => {
           return (
             <MapComponent
               setQuery={setQuery}
-              setPolygon={value}
-              defaultZoom={this.props.defaultZoom}
-              maxZoom={this.props.maxZoom}
-              minZoom={this.props.minZoom}
-              facetData={this.props.facetData}
-              clickIdHandler={this.props.clickIdHandler}
+              value={value}
+              data={data}
             />
           );
         }}
@@ -307,13 +302,13 @@ export default class ReactiveMap extends React.Component {
 }
 
 ReactiveMap.propTypes = {
-  mapComponentId: PropTypes.string.isRequired,
-  clickIdHandler: PropTypes.func.isRequired,
+  componentId: PropTypes.string.isRequired,
+  // clickIdHandler: PropTypes.func.isRequired,
 };
 
 ReactiveMap.defaultProps = {
-  defaultZoom: 6,
+  zoom: 6,
   maxZoom: 10,
   minZoom: 0,
-  facetData: []
+  data: []
 };
