+++ src/index.html
@@ -8,9 +8,7 @@
 
 <body>
   <div class="container">
-    <div id="app">
-      <!-- Where the main React App will be stored -->
-    </div>
+    <div id="app"></div>
   </div>
 </body>
